230221 김범중(2)

+++ AR41DX/Client3D/Include/GameObject/Monster.h
@@ -40,7 +40,7 @@ public:
 		return m_MoveSpeed;
 	}
 
-	float GetMonsterHP()
+	int GetMonsterHP()
 	{
 		return m_MonsterHP;
 	}
